made all build successfully
updated build and test to recognize Dev15Tests

re-arranged test export provider
