+++ BuildAndTest.proj
@@ -79,16 +79,16 @@
 
       <!-- MakeConst unit tests tracked by https://github.com/dotnet/roslyn/issues/5918 -->
       <TestAssemblies Condition="'$(Test64)' != 'true'" 
-                      Include="$(OutputDirectory)\$(IncludePattern)" />
+                      Include="$(OutputDirectory)\$(IncludePattern);$(OutputDirectory)\Dev15Tests\$(IncludePattern)" />
 
       <TestAssemblies Condition="'$(Test64)' == 'true'" 
-                      Include="$(OutputDirectory)\$(IncludePattern)" 
+                      Include="$(OutputDirectory)\$(IncludePattern);$(OutputDirectory)\Dev15Tests\$(IncludePattern)" 
                       Exclude="$(OutputDirectory)\Roslyn.Interactive*" />
 
       <CoreTestAssemblies Include="$(CoreClrTestDirectory)\$(IncludePattern)" />
 
       <MisplacedTestAssemblies Include="$(OutputDirectory)\**\$(IncludePattern)"
-                               Exclude="$(OutputDirectory)\$(IncludePattern);@(CoreTestAssemblies)" />
+                               Exclude="$(OutputDirectory)\$(IncludePattern);$(OutputDirectory)\Dev15Tests\$(IncludePattern);@(CoreTestAssemblies)" />
     </ItemGroup>
 
     <Error Condition="'@(MisplacedTestAssemblies->Count())' &gt; 0"
