+++ src/EditorFeatures/Test/EditorServicesTest.csproj
@@ -218,12 +218,12 @@
     <Compile Include="Extensions\WorkspaceExtensions.cs" />
     <Compile Include="GoToAdjacentMember\AbstractGoToAdjacentMemberTests.cs" />
     <Compile Include="MoveType\AbstractMoveTypeTest.cs" />
+    <Compile Include="Preview\MockPreviewPaneService.cs" />
+    <Compile Include="Preview\TestOnly_CompilerDiagnosticAnalyzerProviderService.cs" />
     <Compile Include="Structure\AbstractSyntaxNodeStructureProviderTests.cs" />
     <Compile Include="Structure\AbstractSyntaxStructureProviderTests.cs" />
     <Compile Include="Structure\AbstractSyntaxTriviaStructureProviderTests.cs" />
     <Compile Include="Structure\BlockStructureServiceTests.cs" />
-    <Compile Include="Preview\MockPreviewPaneService.cs" />
-    <Compile Include="Preview\TestOnly_CompilerDiagnosticAnalyzerProviderService.cs" />
     <Compile Include="DocCommentFormatting\DocCommentFormattingTests.cs" />
     <Compile Include="DocumentationComments\AbstractDocumentationCommentTests.cs" />
     <Compile Include="DocumentationComments\AbstractXmlTagCompletionTests.cs" />
@@ -266,7 +266,6 @@
     <Compile Include="MetadataAsSource\MetadataAsSourceTests.cs" />
     <Compile Include="MetadataAsSource\MetadataAsSourceTests.CSharp.cs" />
     <Compile Include="MetadataAsSource\MetadataAsSourceTests.VisualBasic.cs" />
-    <Compile Include="MinimalTestExportProvider.cs" />
     <Compile Include="NavigateTo\NavigateToTestAggregator.Callback.cs" />
     <Compile Include="NavigateTo\NavigateToTestAggregator.cs" />
     <Compile Include="Structure\BlockSpanTests.cs" />
@@ -287,8 +286,6 @@
     <Compile Include="Tagging\AsynchronousTaggerTests.cs" />
     <Compile Include="Tagging\TagSpanIntervalTreeTests.cs" />
     <Compile Include="TestExportProvider.cs" />
-    <Compile Include="TestExtensionErrorHandler.cs" />
-    <Compile Include="TestOptionsServiceFactory.cs" />
     <Compile Include="TextEditorFactoryExtensions.cs" />
     <Compile Include="TextEditor\TextBufferAssociatedViewServiceTests.cs" />
     <Compile Include="TextEditor\TryGetDocumentTests.cs" />
@@ -302,14 +299,8 @@
     <Compile Include="Utilities\SymbolEquivalenceComparerTests.cs" />
     <Compile Include="Utilities\TestEditorOptions.cs" />
     <Compile Include="Utilities\TestWaitContext.cs" />
-    <Compile Include="Utilities\TestWaitIndicator.cs" />
     <Compile Include="Workspaces\CSharpTestWorkspaceFixture.cs" />
     <Compile Include="Workspaces\MefTestWorkspace.cs" />
-    <Compile Include="Workspaces\NoCompilationConstants.cs" />
-    <Compile Include="Workspaces\NoCompilationContentTypeDefinitions.cs" />
-    <Compile Include="Workspaces\NoCompilationContentTypeLanguageService.cs" />
-    <Compile Include="Workspaces\NoCompilationDocumentDiagnosticAnalyzer.cs" />
-    <Compile Include="Workspaces\NoCompilationLanguageServiceFactory.cs" />
     <Compile Include="OptionExtensions.cs" />
     <Compile Include="Workspaces\TestForegroundNotificationService.cs" />
     <Compile Include="Workspaces\TestAddMetadataReferenceCodeActionOperationFactoryWorkspaceService.cs" />
