+++ src/EditorFeatures/TestUtilities/ServicesTestUtilities.csproj
@@ -29,6 +29,10 @@
     <PlatformTarget>ARM</PlatformTarget>
   </PropertyGroup>
   <ItemGroup>
+    <ProjectReference Include="..\..\Compilers\Core\Portable\CodeAnalysis.csproj">
+      <Project>{1ee8cad3-55f9-4d91-96b2-084641da9a6c}</Project>
+      <Name>CodeAnalysis</Name>
+    </ProjectReference>
     <ProjectReference Include="..\..\Features\Core\Portable\Features.csproj">
       <Project>{edc68a0e-c68d-4a74-91b7-bf38ec909888}</Project>
       <Name>Features</Name>
@@ -52,6 +56,10 @@
     <ProjectReference Include="..\..\Workspaces\Remote\ServiceHub\ServiceHub.csproj">
       <Project>{80fddd00-9393-47f7-8baf-7e87ce011068}</Project>
       <Name>ServiceHub</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\..\Workspaces\VisualBasic\Portable\BasicWorkspace.vbproj">
+      <Project>{57ca988d-f010-4bf2-9a2e-07d6dcd2ff2c}</Project>
+      <Name>BasicWorkspace</Name>
     </ProjectReference>
     <ProjectReference Include="..\Core\EditorFeatures.csproj">
       <Project>{3CDEEAB7-2256-418A-BEB2-620B5CB16302}</Project>
@@ -89,11 +97,12 @@
     <Compile Include="DirectoryExtensions.cs" />
     <Compile Include="ExceptionUtilities.cs" />
     <Compile Include="MinimalTestExportProvider.cs" />
-    <Compile Include="TestExportProvider.cs" />
+    <Compile Include="ServiceTestExportProvider.cs" />
     <Compile Include="TestExtensionErrorHandler.cs" />
     <Compile Include="TestOptionsServiceFactory.cs" />
     <Compile Include="TestWaitIndicator.cs" />
     <Compile Include="Remote\InProcRemostHostClient.cs" />
+    <Compile Include="TestWorkspaceName.cs" />
     <Compile Include="Threading\ConditionalWpfFactAttribute.cs" />
     <Compile Include="Threading\StaTaskScheduler.cs" />
     <Compile Include="Threading\WpfFactAttribute.cs" />
@@ -101,6 +110,11 @@
     <Compile Include="Threading\WpfTestCase.cs" />
     <Compile Include="Traits.cs" />
     <Compile Include="VisualStudioSendKeys.cs" />
+    <Compile Include="Workspaces\NoCompilationConstants.cs" />
+    <Compile Include="Workspaces\NoCompilationContentTypeDefinitions.cs" />
+    <Compile Include="Workspaces\NoCompilationContentTypeLanguageService.cs" />
+    <Compile Include="Workspaces\NoCompilationDocumentDiagnosticAnalyzer.cs" />
+    <Compile Include="Workspaces\NoCompilationLanguageServiceFactory.cs" />
   </ItemGroup>
   <ItemGroup>
     <InternalsVisibleTo Include="Roslyn.VisualStudio.RemoteHostClientMock" />
@@ -121,12 +135,14 @@
     <InternalsVisibleToTest Include="Roslyn.Compilers.VisualBasic.Test.Utilities" />
     <InternalsVisibleToTest Include="Roslyn.InteractiveHost.UnitTests" />
     <InternalsVisibleToTest Include="Roslyn.Services.Editor.UnitTests" />
+    <InternalsVisibleToTest Include="Roslyn.Services.Editor.UnitTests2" />
     <InternalsVisibleToTest Include="Roslyn.VisualStudio.Test.Utilities" />
     <InternalsVisibleToTest Include="Roslyn.VisualStudio.Next.UnitTests" />
   </ItemGroup>
   <ItemGroup>
     <None Include="project.json" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="..\..\..\build\Targets\VSL.Imports.targets" />
   <Import Project="..\..\..\build\Targets\Roslyn.Toolsets.Xunit.targets" />
 </Project>
