+++ src/EditorFeatures/TestUtilities/TestWorkspaceName.cs
@@ -1,0 +1,9 @@
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
+namespace Microsoft.CodeAnalysis.Test.Utilities
+{
+    public static class TestWorkspaceName
+    {
+        public const string Name = "Test";
+    }
+}
