+++ src/VisualStudio/Core/Def/Implementation/Remote/RemoteHostOptions.cs
@@ -8,9 +8,6 @@ namespace Microsoft.VisualStudio.LanguageServices.Remote
     internal static class RemoteHostOptions
     {
         [ExportOption]
-        public static readonly Option<bool> RemoteHostTest = new Option<bool>(OptionName, nameof(RemoteHostTest), defaultValue: false);
-
-        [ExportOption]
         public static readonly Option<bool> RemoteHost = new Option<bool>(nameof(InternalFeatureOnOffOptions), nameof(RemoteHost), defaultValue: true,
             storageLocations: new LocalUserProfileStorageLocation(InternalFeatureOnOffOptions.LocalRegistryPath + nameof(RemoteHost)));
