+++ src/VisualStudio/Core/Test.Next/project.json
@@ -1,9 +1,22 @@
 ﻿{
   "dependencies": {
-    "Newtonsoft.Json": "8.0.3"
+    "Newtonsoft.Json": "8.0.3",
+    "Moq": "4.2.1402.2112",
+    "Microsoft.VisualStudio.Composition": "14.2.19-pre",
+    "Dev15BinariesForRoslyn": "4.0.0.0",
+    "Microsoft.VisualStudio.Imaging": "15.0.25604-Preview4",
+    "Microsoft.VisualStudio.Utilities": "15.0.25604-Preview4",
+    "Microsoft.VisualStudio.ImageCatalog": "15.0.25604-Preview4",
+    "Microsoft.VisualStudio.Text.UI": "15.0.25604-Preview4",
+    "Microsoft.VisualStudio.CoreUtility": "15.0.25604-Preview4",
+    "Microsoft.VisualStudio.Language.Intellisense": "15.0.25604-Preview4",
+    "Microsoft.VisualStudio.Text.Data": "15.0.25604-Preview4",
+    "Microsoft.VisualStudio.Text.Logic": "15.0.25604-Preview4",
+    "Microsoft.VisualStudio.Text.UI.Wpf": "15.0.25604-Preview4",
+    "Microsoft.VisualStudio.Editor": "15.0.25604-Preview4",
   },
   "frameworks": {
-    "net46": {}
+    "net46": { }
   },
   "runtimes": {
     "win7": { }
