+++ src/Workspaces/VisualBasic/Portable/BasicWorkspace.vbproj
@@ -53,6 +53,7 @@
     <InternalsVisibleToTest Include="Roslyn.Services.UnitTests" />
     <InternalsVisibleToTest Include="Roslyn.Services.VisualBasic.UnitTests" />
     <InternalsVisibleToTest Include="Roslyn.VisualStudio.Services.UnitTests" />
+    <InternalsVisibleToTest Include="Roslyn.Services.Test.Utilities" />
   </ItemGroup>
   <ItemGroup>
     <Compile Include="CaseCorrection\VisualBasicCaseCorrectionService.Rewriter.vb" />
